requirements.txt
requests>=2.31.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytest>=7.4.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.0.0
pyodbc
panda
openpyxl